% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_loss.R
\name{head_loss}
\alias{head_loss}
\title{Head loss over the length of pipe}
\usage{
head_loss(d, q, l, rc = 1e-06, v = 1.01e-06, g = 9.81, x1 = 0.06)
}
\arguments{
\item{d}{Diameter in meters}

\item{q}{Flow rate in cubic meters per second}

\item{l}{Length of pipe in meters}

\item{rc}{Roughness coefficient. Absolute roughness (in meters) for
Universal equation.}

\item{v}{Kinematic viscosity of fluid in square meters per second. By
default use the value for water at 20 Celsius degree
(\code{v = 1.01e-6 m^2/s}). Unnecessary for empirical
equations.}

\item{g}{Gravitational acceleration. By default use the value
\code{g = 9.81 m/s^2}}

\item{x1}{Initial parameter of f for Newton-Raphson.
By default \code{x1=0.06}.
Unnecessary for empirical or explicit equations.}
}
\value{
hf Head loss in meters
}
\description{
\code{headLoss} calculate the head loss by friction in pipes by different
methods
}
\examples{
head_loss(d = 0.025, q = 31e-6, l = 100, rc = 1e-4)
}
