% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm.R
\name{hm}
\alias{hm}
\title{Função hm}
\usage{
hm(hr, hs, dr_com, q, lr, per, rc, ds_com, ls, pes)
}
\arguments{
\item{hr}{Altura de recalque}

\item{hs}{Altura de sucção}

\item{dr_com}{Diâmetro comercial}

\item{q}{Vazão}

\item{lr}{Comprimento do recalque}

\item{per}{Comprimento equivalente linha de recalque}

\item{rc}{Coficiente de rugosidade}

\item{ds_com}{diametro linha de succao}

\item{ls}{comprimento da linha de succao}

\item{pes}{Comprimento equivalente linha de succao}
}
\value{
numero
}
\description{
Função hm
}
