% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmpump.R
\name{hmpump}
\alias{hmpump}
\title{Title}
\usage{
hmpump(q_bomba, hm_bomba)
}
\arguments{
\item{q_bomba}{Vazoes da motobomba}

\item{hm_bomba}{Altura manometrica da motobomba}
}
\value{
Um data.frame
}
\description{
Title
}
\examples{
Q_BC21_15 <- c(19.2, 18.2, 17.2, 16.0, 13.3, 9.9)
Hm_BC21_15 <- c(9, 10, 11, 12, 14, 16)
hmpump(Q_BC21_15, Hm_BC21_15)
}
