% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmsis.R
\name{hmsis}
\alias{hmsis}
\title{Title}
\usage{
hmsis(hr, hs, dr_com, q, lr, per, ds_com, ls, pes)
}
\arguments{
\item{hr}{Altura de recalque}

\item{hs}{Altura de sucção}

\item{dr_com}{Diâmetro comercial}

\item{q}{Vazão}

\item{lr}{Comprimento do recalque}

\item{per}{Comprimento equivalente linha de recalque}

\item{ds_com}{diametro linha de succao}

\item{ls}{comprimento da linha de succao}

\item{pes}{Comprimento equivalente linha de succao}
}
\value{
um data.frame
}
\description{
Title
}
\examples{
hmsis(
  hr = 20, hs = 1, dr_com = 60 / 1000, q = 10 / 3600, lr = 50, per = 148,
  ds_com = 75 / 1000, ls = 2, pes = 305
  )
}
