% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pump.R
\name{plot_pump}
\alias{plot_pump}
\title{Title}
\usage{
plot_pump(df)
}
\arguments{
\item{df}{Objeto resultado da funcao pumpsis}
}
\value{
Um Plot da curva caracteristica da motombomba
}
\description{
Title
}
\examples{
Q_BC21_15 <- c(19.2, 18.2, 17.2, 16.0, 13.3, 9.9)/3600
Hm_BC21_15 <- c(9, 10, 11, 12, 14, 16)
curva_bomba <- hmpump(Q_BC21_15, Hm_BC21_15)
plot_pump(curva_bomba)
}
