% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sys.R
\name{plot_sys}
\alias{plot_sys}
\title{Title}
\usage{
plot_sys(df)
}
\arguments{
\item{df}{Objeto resultado da funcao hmsis}
}
\value{
Um Plot da curva caracteristica do sistema
}
\description{
Title
}
\examples{
curva_sistema <- hmsis(
  hr = 20, hs = 1, dr_com = 60 / 1000, q = 10 / 3600, lr = 50, per = 148,
  ds_com = 75 / 1000, ls = 2, pes = 305
)
plot_sys(curva_sistema)
}
